+++ src/libraries/vup/GPU_Storage/Element_buffer.cpp
@@ -14,7 +14,7 @@ vup::Element_buffer::Element_buffer(const gl::storage flags)
 vup::Element_buffer::Element_buffer(const std::vector<unsigned int>& indices,
                                     const gl::storage flags)
     : Buffer(GL_ELEMENT_ARRAY_BUFFER, indices, flags) {
-    m_count = indices.size();
+    m_count = static_cast<unsigned int>(indices.size());
 }
 
 void vup::Element_buffer::set_data(const std::vector<unsigned int>& data) {
